Removes places where we were filtering for the gprd environment.
This removes many places where we were filtering for environment,
because the alartmanager will not page on the staging environment
it makes more sense to alert on everything equally.  There may be some
extra noise after this merge but I suggest we see what happens and
address alerts individually as they appear in the staging environment.