Ignore workhorse git http sessions for low-volume environments such as canary.
In order for range to be considered high, it needs to be at least 3
sigma out of normal range, but also, it needs to be at least 100 sessions
open.

This will preclude low volume environments such as canary from firing
on this.

See https://prometheus.io/blog/2015/06/18/practical-anomaly-detection/
for details of this technique.