WIP: restructuring README to foucs on On-Call
The README provided an outline with general
guidance for incident response. But, there
were conflicting pages with duplicate information
pertaining to the roles and responsibilities
shared during incident management.

I consolidated into the README a couple pages
that were under the `howto/` section of the
project. And I've added an `on-call/` section
that will include the checklists to follow
at the start of both shifts and incidents.

fixes gitlab-com/gl-infra/infrastructure#6765