Adds long-term saturation forecasting
Long-term saturation forecast is not focused on immediate changes, but
rather long-term trends. It uses week-on-week growth to predict
several metrics:

Average Values:

1. What the will the average value of the saturation metric be in two
   weeks time?

2. What the will the average value of the saturation metric be in 30
   days time?

Sapdex Trends:

Averages don't work very well for saturation as it can be spiky.
A second measure we monitor is how much time a metric spends outside of
it's hard and soft SLO values. The soft SLO is a region in which the SLO
will not alert, but is considered a warning. The hard SLO will generate
an alert.

Using a method similar to apdex, a metric will score 0.5 when it exceeds
its soft SLO and 0 when it exceeds its hard SLO.

Using these values we can then calculate :

1. What will the Sapdex score for the component be in 30 days time?