Remove patroni from the single_threaded_cpu saturation metric
While we need to somehow see if pgbouncer on patroni nodes is saturating a CPU, this rule currently
captures (and results in alerts for) anything on a patroni node using most of a single CPU.  This
includes postgres itself, doing normal operations including backups.  For this to be of value
we need to find a way to extract how busy pgbouncer is specifically.