Increase WALE backup failure threshold to 2 consecutive failures
The previous 48 hour limit combined with the fact that our backups are
now sizeable enough to take the better part of a day and _maybe_ go over
means we might get paged unnecessarily. Two consecutive failures (72h)
should be a sign that a problem worth the attention of a DBRE is indeed
going on.