Do not index a malformed field in ES
This is found in the GKE indexes. It is emitted by the Docker registry,
and is sometimes a string, sometimes an object. By not indexing it, we
will not be able to search on it, but all logs should end up in ES
instead of dropping some of them based on type.