Split index mappings
Instead of having one mapping for all indexes. This will allow us to
support field names of different types across indexes, and improve
signal:noise in each individual index mapping (as 100% of it will be
relevant).

This breaks the ability to configure different mappings in prod and
nonprod, but reduces the need for copy paste between these two
environments.