Adjust logging alerts
Express dropped log alert as a ratio, and alert at 0.5% instead of
non-zero.

Nudge pubsubbeat error alert up from zero to 1 event every 10 seconds.
Note that dropped logs are warnings, not errors, from pubsubbeat.

Nudge pubsubbeat warning alert up from zero to 1 event every 2 seconds.
Warnings are highly correlated with dropped logs, but since this one
can't be expressed as a ratio, it is more likely to fire on pubsubbeats
with higher traffic. This risks it firing due to a dropped log ratio
that is too low to justify alerting. In the future, we may want to dial
up this threshold even higher, or eliminate the alert.