Add instructions to `k8s-gitlab.md` for working with the new `gitlab-helmfiles` repo
As we are in the process of migrating and merging repos under `k8s-workloads` into
a single `gitlab-helmfiles` repo, we need to update this document to reference
its inclusion and the steps to use it.

This is an initial addition, and will be further refactored as more workloads
are migrated.