Add recording rules for sidekiq jobs with no urgency
mailers is a special queue that is built into rails, and has no 'urgency' label, meaning it is not picked up by the existing recording rules.  This adds rules explicitly for the empty/non-existent urgency.

See also https://gitlab.com/gitlab-com/gl-infra/production/-/issues/2062 for where we were not alerted to an issue because of this lack