Add a disableOpsRatePrediction option for metrics catalog services
Some services, notably monitoring, don't have a normally distributed
operation rate. This means that we get unneccessary alerts for this
service as traffic spikes (without any other apparent negative effect)

For this reason, we disable this for the monitoring service.

Additionally, this MR removes the alert from prometheus, and relies
exclusively on the Thanos version.