Add apdex weight scores to allow us to aggregate apdex scores across multiple Prometheus servers in Thanos.
This will help with the migration to sidekiq k8s.

Part of https://gitlab.com/gitlab-com/gl-infra/scalability/-/issues/351

Additionally, by providing a weight, will help us answer this question
from @sloyd:

>>>
Is this issue contributing to the continuing problems we see with
Sidekiq availability performance (in addition to authorized_projects)?
If so that might be another reason to escalate this. We're continuing to
see sidekiq drag on the SLA board even after the improvements end of
last month.
https://dashboards.gitlab.net/d/general-slas/general-slas?orgId=1&from=1588291200000&to=1588809599000
>>>

--- https://gitlab.com/gitlab-org/gitlab/-/issues/216458#note_336730395