Include the average case apdex and error rate scores
Currently we display the worst case over the interval time period
for each of these metrics.

For short ranges (eg 6hrs) this works well at picking up anomalies that
would otherwise be erased by an `avg_over_time`.

However, when we zoom out to  periods of weeks, this can become extremely
distorted. If a user is unaware of this trade-off, they may conclude
that the metrics are worst than they actually are.

This change adds an average value alongside the worst case.

This makes it easier to spot over longer periods what the SLI is doing.