Remove env labels from general Cloudflare metrics
No longer needed, now that we scrape zones from different instances of
the Cloudflare exporter, deployed in the appropriate env (and therefore
with the appropriate external labels).

Pager routing is done on `env` (the external label) and not
`environment` (the previously-overridden, non-external label), so it's
no surprise that we saw pages for staging zones - but it is a bit of a
surprise that these alerts had `environment: gprd`. So it's possible the
staticLabels config wasn't even working as intended, but that's
orthogonal to this change and should be investigated separately.