Allow different burn rate periods to be specified for different resources
Currently, all saturation points are measured on a minute-by-minute
basic.

This does not necessarily make sense. For example, disk throughput
metrics define a maximum sustained throughput, not an absolute limit.

This change adds an optional `burnRatePeriod` attribute for saturation
resources. This is used in the `rangeInterval` value, and also used as
the minimum interval on detail graphs.

This should hopefully ensure that our alerting on these resources is
more stable.