Fix Sidekiq shard-level throughput
For some reason, shard wasn't excluded from sidekiq_jobs_completion_rate
as it was from sidekiq_jobs_completion_seconds and sidekiq_jobs_failed.

This strengthens the case for these to be in the metrics catalog, but
for now, we should at least fix the rule so the Sidekiq shard dashboard
can show throughput again.