Slacken time for pullmirror alert
The jobs related to pullmirror are considered non-latency-sensitive, and
as such are run on our sidekiq catchall fleet. This fleet is allowed to
become saturated on compute resources more than others, and job
latencies can increase as a result. Pullmirror delays are therefore not
necessarily actionable, and we should only alert on truly long-lived
pile-ups.