Add Global search initial indexing queue size to sidekiq overview
This is a new metric that was introduced in
https://gitlab.com/gitlab-org/gitlab/-/merge_requests/33384 where more
sidekiq Elasticsearch workloads were moved over to using a custom queue
mechanism. This chart is exactly the same as the "Global search
incremental indexing queue length" one adjecent to it in this dashboard
except it is a different queue that is only responsible for newly
created projects or newly indexed projects where the other one is
tracking incremental updates to already indexed projects.

I did consider adding this line to the existing chart but it seems
like there is no good way to do that given the charts both collect an
identical metrics from every rails node scraped by Prometheus meaning
that there are already many identical lines being displayed in each
graph. If we had both in the same graph it would be very difficult to
filter to just one of them or visually distinguish them with all the
noise. We did discuss aggregation approaches in
https://gitlab.com/gitlab-com/runbooks/-/merge_requests/2046 but still
did not land on anything due to there being no way to know which is the
most up to date data point for the queue size so instead we display all
the lines from every node.