Super simple viewer of the redis flamegraphs
Collected as a result of work in https://gitlab.com/gitlab-com/gl-infra/scalability/-/issues/321.  Assumes the user of the browser has basic read access to the GCS buckets (not sure if this is wide-spread, but we'll find out.

Not intended to be a polished thing, just an initial way to view the data