Reuse the apdex weight calculation result in the apdex score calculation
Currently, we count the number of requests in the apdex score twice.

One for the weight calculation, and again for the denominator on the
apdex score calculation.

This change optimises the query, substituting the recording rule from
the weight calculation into the apdex score calculation.

In adhoc testing this halves the latency of apdex calculation.