Update k8s Alertmanager config in CI
The alertmanager jsonnet generation step already outputs 2 files: an
alertmanager config, and a helm value snippet. Both are uploaded to GCS.
The helm value snippet contains the alertmanager config file nested as a
field, and is consumed by gitlab-helmfiles pipelines, which insert
templates and create a secret which is passed to the Alertmanager CRD.
The raw config file is consumed by chef on GCE alertmanager nodes.

This commit adds a third file: a kubernetes secret manifest. It is
identical to what helm produces already.
https://gitlab.com/gitlab-com/gl-infra/k8s-workloads/gitlab-helmfiles/-/merge_requests/136
changes our Alertmanager k8s setup to not create the secret object, but
to retrieve it by name, with a name matching the manifest that this
pipeline will now create.

The CI job in this repo that generates the Alertmanager config now runs
in the ops CI environment. The environment-agnostic SERVICE_KEY
configured in ops.gitlab.net is already the same service account key
used by the ops environment, so we wouldn't expect the GCS access to be
changed. Using environment-scoped CI variables should allow our pipeline
to access the Kubernetes API of the ops cluster, allowing the pipeline
to `kubectl apply` the secret directly.

After both MRs are deployed and working, we should be able to stop
uploading the helm value snippet, as it will no longer be used. Once GCE
alertmanagers are turned down, we can also stop uploading the raw config
file. Both of these considerations are written into the jsonnet as
comments.

Add "thanos-rule" to the dummy secrets snitch environment list, so that
locally generated config more closely matches what's in the bucket, to
assist testing.