Remove Gitaly label from deployment alerts
Gitaly gets the alerts duplicated in the #gitaly-alerts channel in
Slack when the team is set to Gitaly. This means that for each deploy,
the channel gets an alert about a current deploy. This adds noise, as
we deploy often. 

This change removes the team, and thus the duplicate alert.