Do not treat 0 primaries as split-brain
The case of 0 primaries should trigger a different alert
than the split-brain case.  Those two scenarios require
different handling, and the 0 primaries case may briefly
trigger during healthy failover events.  Mischaracterizing
those events as split-brain could lead the on-call engineer
to needless anxiety and effort.

Also, now that we differentiate this alert by "type" label,
that label needs to be added to the promql queries in this
alert's "description" and "title" attributes.