Fail pipeline when duplicate dashboards are uploaded
It's possible for two dashboards in the same folder to have the same
title, and so end up with the same URL in Grafana. If that happens, the
second one uploaded will 'win' every time, and the first dashboard won't
be available anywhere.

It's hard to spot this so we add it to CI. Unfortunately the URL is
generated by Grafana so we can only check this when the dashboards are
being deployed, which is after the merge request will have been merged,
but it's better than nothing.