Remove the expiry time from SSL alerts to issues
The gitlab webhook integration uses the "title" as an identification key, so including the days-left-to-expiry in the title results in a new alert + issue every time.  This moves that out of the title (making it the same for any given monitored endpoint), into an annotation, where it should show up in the alert message/issue.