Allow alerts to go to both pager and issue tracker
We do not want paging-ness and issue raising to be mutually exclusive:
issue trackers are not a paging mechanism.

Implemented by removing `pager: issue` from the issue-tracker routing
logic, and solely using the `incident_project` label to route alerts to
issue trackers. This Alertmanager route already had `continue: true`
statically configured, and this commit inlines the variable.

The `incident_project` label used to be named `project`, but we've
renamed it to make its purpose clearer. This change was made in both the
routing tree and the alerts under `rules` that used it simultaneously.

Remove redundant `pager` label from alerts that had `pager: issue`.

Fix an alert label test whose description asserted that all alerts with
severity of 2 or greater should page to actually express that, and
downgrade severities of the current issue alerts (none of which page) to
satisfy the test.