Add a job a triggerable job to verify stage groups
This is an extra job that can be manually triggered on merge requests
or when $CHECK_STAGE_GROUPS is set to "1". The job will also run when
the mapping file contains changes

To do this, a lot of the jobs that were using `except: schedules` have
been converted to a shared "default-job-rules" configuration that can
be extended.

This way, we can make sure that we can trigger this single job from
the Scalability pipeline