Add a job to verify stage groups mapping
This is an extra job that can be manually triggered on merge requests.
When $CHECK_STAGE_GROUPS is set to "1" or the mapping file has
changed, the job will run automatically.

To do this, a lot of the jobs that were using `except: schedules` have
been converted to a shared "default-job-rules" configuration that can
be extended.

This way, we can also make sure that this is job can be triggered
exclusively from a Scalability pipeline.