Rework job rules
This is so we can run a single job triggered by an upstream pipeline.

This adds 2 commonly used anchors for avoiding schedules and the stage
group check. These should be included in most jobs.

By default test jobs should extend the `default-job-rules`. This will
makes sure it doesn't run for schedules, or when we triggered a
pipeline for a single job.

But if a job defines it's own rules, those take precedence over all
rules defined in `default-job-rules. In this case we should consider
adding the appropriate anchors to the jobs.