SLIs can only be owned by one team
We don't actually need to support more than one team per service yet,
and it's not clear that we will have to in the future. This is the case
because teams in the metrics-catalog don't necessarily map 1:1 with
GitLab company teams. They usually do, but there are virtual teams, such
as workhorse, to route alert notifications to single places that are
monitored by multiple "real" teams.

In all SLIs that contained more than 1 team, there were 2 teams, one of
which was "workhorse". These have been changed to just workhorse, after
a discussion in the MR containing this commit.