Alert as soon as any alertmanager failure detected
By removing `for` clause.

This is not a foolproof way of ensuring that our alerting routes are
working. If an integration is misconfigured (e.g. due to a broken
template) then those alerts will fail - and if the route pertaining to
AlertmanagerNotificationsFailing is broken, we won't see that! A big
improvment would be to use a dead man's switch (e.g. Dead Man's Snitch)
to catch broken integrations in our paging path, as a fallback. Of
course if our paging service (today, Pagerduty) is having an outage,
then our (paging) alerting will be down, and we'd have to rely on
spotting other channels for these snitches (e.g. Slack).