Change asdf plugin used for jsonnet-bundler
This one (https://github.com/beardix/asdf-jb) is in the official
repositories and so is installable with `asdf plugin add jb`, with no
repo qualifier, and is discoverable with `asdf plugin list-all`. This
makes the comment in .tool-versions true, as I had forgotten that
`jsonnet-bundler` was outside the official repos
(https://github.com/trotttrotttrott/asdf-jsonnet-bundler).

See https://github.com/trotttrotttrott/asdf-jsonnet-bundler/issues/1.