Reduce the error ratio SLO for Gitaly nodes
During the end-of-year holidays, the error ratios are higher than normal
for several Gitaly nodes.  This is due to a combination of lower overall
request rate and a few git repos having large spikes in their rate of
canceled gRPC calls.  The SLO currently counts Canceled gRPC calls as
errors, so this spikes the ratio of errors to overall requests.

See here for more details, background, and links to example incidents:
https://gitlab.com/gitlab-com/gl-infra/infrastructure/-/issues/12220