metrics-catalog: rename kubeDeployments
To kubeResources. We currently use this field to populate dashboards
with container metrics charts (e.g. CPU and memory). Examples:

- https://dashboards.gitlab.net/d/mailroom-main/mailroom-overview?orgId=1&viewPanel=16
- https://dashboards.gitlab.net/d/mailroom-kube-deployments/mailroom-kube-deployment-detail?orgId=1&from=now-6h%2Fm&to=now%2Fm&var-PROMETHEUS_DS=Global&var-environment=gprd&var-stage=main

The "deployment" in "kubeDeployments" didn't refer to a capital-D
Kubernetes Deployment, but to a logical grouping of instances that
happen to be containerized processes, that provide the service. These
containers can be controlled by a Deployment, a StatefulSet, or
something else entirely. At this level, we don't care.

The recording rules that populate these dashboards are derived by
adding a "deployment" label that is equal in value to the "deployment"
label on the container's pod. Our code that generates Kubernetes
resources (helm, tanka) adds this label based on the name of
Deployments, but also of StatefulSets. It's more hassle than it's worth
to change this label's name at this time.

Add a "kind" field to kubeResources to capture the metadata of what the
controller type is (e.g. "Deployment"), but we don't use this anywhere
just yet.