Fix pre-mature exit in storage_rebalance.rb
With each project we migrate, we shrink the set we're selecting from (as we're
selecting by repository_storage) while advancing the pagination index forward
at a constance pace. At some point we're going request a page that's outside
the shrunk set, getting empty list in response. Here we do a start-over from page
1 whenever we think we reached the end of the set, and only exiting when we get two
empty responses in a row.