Split PostUploadPack into own SLO
We often see alerts for gitaly/goserver apdex SLO violations, e.g.
https://gitlab.com/gitlab-com/gl-infra/production/-/issues/3316 and
https://gitlab.com/gitlab-com/gl-infra/production/-/issues/3279. Often,
the culprit is slow PostUploadPack requests.

The metrics-catalog SLI description for gitaly/goserver says "The apdex
score is based on a subset of GRPC methods which are expected to be
fast". We can't reasonably expect all fetches or clones to be equally
fast: it's dependent on the size of the repository and of the objects
being fetched.