Open incident issues for metamon alerts
Add the production project to all s4 and s3 meta-monitoring alerts. S2+
already go to Pagerduty.

Related: https://gitlab.com/gitlab-com/gl-infra/infrastructure/-/issues/12328

This runs the risk of being noisy, because the `#alerts` channel is
typically not monitored as rigorously as `#alerts-general`, or of course
Pagerduty. Alerts that only go there may flap regularly. We should not
hesitate to use the silencer and triage that alert in an issue for
removal / rewriting.