Update documentation on updating a Kubernetes cluster
As part of https://gitlab.com/gitlab-com/gl-infra/infrastructure/-/issues/12367
it was discovered that with the latest state of GKE, these instructions are no
longer workable. They have now been re-written to accomdate the best solution
for doing upgrades while still keeping as much information as we can in terraform.