Deduplicate prom rule evaluation failure alerts
The ThanosRuleHighRuleEvaluationFailures alert was really a special case
of PrometheusRuleEvalFailures by job.

Rewrite PrometheusRuleEvalFailures to aggregate by rule_group, to avoid
spamming the incident issue tracker with issues when there is general
rule unavailability, e.g. because thanos-query is unavailable to the
rule engine. Change the description to direct the reader to a
de-aggregated thanos query to dig into the problem.

Related: https://gitlab.com/gitlab-com/gl-infra/infrastructure/-/issues/12689