Reduce apdex threshold for CI runners
Context:
https://gitlab.com/gitlab-com/gl-infra/production/-/issues/3800,
https://gitlab.com/groups/gitlab-org/-/epics/5434

We are currently experiencing pressure on the CI runners service in a
few ways. A rapid action is underway to improve performance, and we are
undertaking some mitigations in the related incident issue. Retreating
from 97% to 95% allows us to spot further regressions, but the current
pager load is broadly unactionable (in a short-term sense).