Use thread-based CPU metrics for redis primary cpu saturation metric
With the addition of io-threading in redis, the total CPU reported by the redis
INFO command (as exported by redis_exporter) is not the correct saturation
metric as it includes IO threads in the total (and these can be *very* busy,
summing to over 1 CPU with the core thread).

This changes the saturation metric to use namedprocess_namegroup_thread_cpu_seconds_total
from process exporter, and only looking at the "redis-server" thread
which is the core (and saturatable) thread, thus excluding the new io_thd_X
threads.

Depends on https://ops.gitlab.net/gitlab-cookbooks/chef-repo/-/merge_requests/5149
to have been deployed before this can be merged, otherwise we'll silently
lose alerting on CPU saturation for the cache and sidekiq redis clusters.