Add a simple symptom ElasicSearch queue alert
This adds an alert for the ElasticSearch queue size.

The fixed value of 50k is picked based on the past 4 weeks. We had an
incident where the queue jumped up to over 100k, but otherwise we
haven't peaked over 50k. So I think that value should be good for
now.

In order to be able to link to a grafana panel from the alert, I've
added it to the Global Search stage group dashboard.