Extract all manifest routes
This extracts all manifest operations into their own SLIs: Split up
into a tighter one for reads (GET or HEAD) requests, and a looser one
for writes (PUT or DELETE).

The SLI for writes also includes the unused PATCH or POST
methods. This ensures we don't miss any methods when excluding a route
from the main apdex.

Now that the fast GET and HEAD operations are extracted out of the
main server SLI, we had to loosen the thresholds there.

It adds a way to make this easier to do in the future, ensuring we
don't miss any method-route combinations. Any route that has a custom
SLI is automatically excluded from the main server apdex.