Fix postgres split brain alert
We currently have a standby PG12 cluster, whose future primary is
replicating from one of our main cluster's replicas. Due to poor service
/ label organisation, we don't have a neat way of separating these 2
clusters with labels, so must use this FQDN hack.

It is _very_ important that this change be reverted when postgres 12 is
rolled out, otherwise we'll be blind to real split-brain scenarios. Add
this MR, and a revert for it, to whatever change issues are dealing with
this.