Add a feature_category label to source SLIs
This adds a feature category label to the recordings for SLIs that hat
a `featureCategory` different from `not_owned`.

We're excluding `not_owned` here because for some SLIs this means the
originating metrics contain their own `feature_category`. This is the
case for Puma and Sidekiq, which emit metrics for different feature
categories. By excluding the feature category here, we're making sure
that the metrics aren't combined the already existing `not_owned`
metrics coming from those services.

We're also adding validation to the SLI definitions. If we add a
`featureCategory` to an SLI, it needs to be one we know of in the
mapping.