Record feature category metrics for SLIs
This adds a feature category recording with a static label for all
SLIs that had a feature category defined.

The recordings are done separately  from the key metrics to avoid
duplication during a transition. This also means we don't need to
adjust the anomaly detection alerts. The downside is that we're
effectively evaluating the same rule twice but but recorded with
different static labels

We're excluding the ones that had the feature category set to
not_owned. Those recordings don't add a lot of value, and some
services like Sidekiq and Puma might have their own feature category
metrics.