Add worker label to sidekiq queue aggregations
To support the ability to have one-queue-per-shard rather than the current
one queue per worker (specifically https://gitlab.com/groups/gitlab-com/gl-infra/-/epics/447)
we need the sidekiq metrics/aggregations to have the worker label as well as queue.

This is acceptable for cardinality because the current mapping is 1:1 on queue:worker
so we'll have exactly the same number of metrics after applying this, they'll just
have an additional label.

Specific issue with justification/discussion is https://gitlab.com/gitlab-com/gl-infra/scalability/-/issues/1014