Refactor the stage group dashboard test
This refactors the test output is more comprehensible on failures.

Before the output would be the JSON-string of the entire dashboard. By
switching to the standard `expect` & `actual` specifiers, we'll get
the actual and expected output in the failure message. Which makes it
slightly easier to debug.

This also splits up some tests so we don't test the same thing in each
one.