Add mechanical sympathy watch for redis duration
This adds a watch for the the p95 redis_duration_s field.

This fields represents the sum of all Redis calls to any redis
instance.

The threshold is currently set to our `slowlog-log-slower-than`
threshold. But these could consist of multiple redis calls to mutliple
redis instances.

Still, if the total time spent talking to several Redis instances
consists of more than 0.01 second, that seems like a lot. The watch is
only going to log the top 5 offenders. So if we notice that the p95
for these is too close to the threshold, we could consider increasing
it.