refactor(ci-runners): use concurrent saturation for srm
Background:
---

Use `gitlab_runner_concurrent` for saturation calculation rather then
`gitlab_runner_limit` for `srmX`. `concurrent` defines how many jobs each runner
manager can run. `limit` is specific for each `[[runner]]` registered on
that runner manager, and also is used to handle the number of Idle
machines. In
https://gitlab.com/gitlab-com/gl-infra/infrastructure/-/issues/13277#note_572139467
we are discussing to increase the number of `limit` to not match
`concurrent` so that we have Idle machines when we are at capacity.

Done:
---

Igonore the `private` and `shared-gitlab-org` shards because those have
multiple `[[runner]]` sections registered and `limit` is the only way to
get saturation per worker.

reference https://gitlab.com/gitlab-com/gl-infra/infrastructure/-/issues/13277