Extract error budget queries into library
This splits up the error budget file that used to contain everything
related to building the error budget in Grafana in a single file.

This is now 3 files:

*libsonnet/stage-groups/error_budget.libsonnet*

This is the file that is now the entry-point for everything using the
error budgets. It contains the configuration for the budgets, and it
exposes the panels to be reused on dashboards.

It exposes the configuration that currently consists of the
availability target (based on our overall SLA), and the
range (28d). Based on SLA and the passed range it will calculate the
budget in seconds.

In the future this will also expose the queries file, to be reused
outside of the dashboards

*libsonnet/stage-groups/panels.libsonnet*

This builds all the panels to display in Grafana. This still accepts a
range, to be able to display the panels for a different range if
needed.

*libsonnet/stage-groups/queries.libsonnet*

This contains the queries for getting the error budget. It allows
passing selectors and aggregationlabels.

This has been extracted to be able to reuse it to get data into
Sisense.