Add initial version of `ops-runner-hit-gitlab` ES watcher.
Part of https://gitlab.com/gitlab-com/gl-infra/delivery/-/issues/1741

This adds a watcher which looks for API/WEB requests to gitlab.com
from the IP addresses of any of our runners we use for deploy/release
jobs in ops.gitlab.net.

If it identifies any access, it puts a slack alert in the channel
ops-runner-gitlab-alert . This will allow us to start seeing when that
alert fires, investigate, and figure out if our tooling needs to be changed
to stop that specific request, or if it's ok and non-critical (in which case
the traffic pattern should be added as an exclusion to this watcher).

As we further refine this alert, we will change the notification strategy
of it to something more prominent.