Add the trailer to sat alert dashboards to get stable Ids
Alerts generated for component saturation assume stable ids for panels (given to Slackline to be able to render the panel and put it in Slack).  These are only applied by the trailer() function, so without this, Slackline tries to render images of non-existent panel ids, which timeout, lead to Grafana errors, and eventually paging if there's enough of them.

See https://gitlab.com/gitlab-com/gl-infra/production/-/issues/4762.