Fine-tune the 'errors' filter for ci runners
Per https://gitlab.com/gitlab-com/gl-infra/production/-/issues/4770#note_588358920 when linking to the runners index for a failure, the metrics that caused it will be a specific subset of failures that have this full text as the prefix.  By starting from that we end up at the real causative errors rather than a firehose of other 'failed' message which are not what triggered alerts.