Rename metric to match new exporter version
When we upgrade the gitlab-exporter for sidekiq-redis, the new exporter
version renames 4 metrics.  1 of those metrics is referenced
by the dashboard and recording rule definitions here in the runbooks repo:

  "sidekiq_queue_latency" renamed to "sidekiq_queue_latency_seconds"

We should merge and apply this runbooks change at roughly the
same time as we upgrade the gitlab-exporter for sidekiq-redis, so that
we have an acceptably small monitoring gap.

Issue: https://gitlab.com/gitlab-com/gl-infra/scalability/-/issues/1029#note_591568530