Use recordings to get avg SQL count
This uses recordings from the rails recording rules to get SQL
transaction count per request.

The as a result the queries will be faster and will have the same
labels. Which would solve the missing data on some dashboards.

Because both recordings are in a different group, it does still suffer
from the possiblity of having more samples in one series compared to
the other. But that's not worse than it is before.