Use new Grafana table for error budget attribution
By default Grafonnet tables will be rendered in an old style table,
using the `type:
'table-old'` (https://github.com/grafana/grafonnet-lib/issues/240).

This happens by passing an array into the styles attribute. When the
`styles` attribute is present, Grafana will treat it as an old table.

However, we were using styles to modify and hide columns:
https://github.com/grafana/grafonnet-lib/blob/3082bfca110166cd69533fa3c0875fdb1b68c329/grafonnet/table_panel.libsonnet
In new tables this needs to be done using transformations.

Sorting using the `sort` attribute is also not supported. So we're
providing this in the default `options` object.