Allow specifying a single jsonnet file for testing
Since we start to have more jsonnet unit tests, running them becomes a
bit slower.

This updates the script to specify a path to run a single one of them
rather than all of them.

Running the script for a single file looks like this:

  → scripts/jsonnet_test.sh periodic-thanos-queries/periodic-query_test.jsonnet
  # periodic-thanos-queries/periodic-query_test.jsonnet
  {
     "verify": "Passed 1 test cases"
  }

Running the script without arguments will still run all of the found
`_test.jsonnet` files.