Show a single color for error budget in minutes
This graph shows a dynamic range: from the begining of last month to
yesterday. This means that error budget in minutes also changed
depending on $__range_ms. Since we cannot perform calculations in the
threshold specification of a stat-panel we don't know the the budget
in minutes to deduct thresholds.

To avoid confusion, we'll show a single color on the foreground.

https://github.com/grafana/grafana/issues/922

This also switches the ordering of the columns so it's visible that
the budget spend in minutes relates to the percentage