ci-runners: add new private shard subnetwork
In gitlab-com/gl-infra/infrastructure#13405
new set of runner managers for the `private` shard has been created. A
new subnetwork was created inside of `ci` VPC for the new ephemeral
runners.

This was preferred over creating VPC peering between the new VPC and the
`default` VPC`. We wanted to use the `ci` VPC because we want these
runner managers to only be assessable by the bastion.
f5ce469c