Fail when jsonnet dashboard compilation fails
In 22bd8f350879dd04e7682921a864e737c3134bc4 we've changed the way the
JSON is generated using a subshell to assign it to a variable using
substitution.

The side effect was that a compilation failure during `./upload.sh` or
`/test-board.sh` would not result in a failure of the script.

Enabling inherit_errexit will cause subshells to fail:

> If set, command substitution inherits the value of the errexit
> option, instead of unsetting it in the subshell environment. This
> option is enabled when POSIX mode is enabled.

https://www.gnu.org/software/bash/manual/html_node/The-Shopt-Builtin.html