Use worker name as label for vector(0) in worker queue length panel
The vector(0) was added to generate a zero if there were no queued jobs for a given worker, rather than having a blank graph with potentially "No data", which looks like a data problem, which could be utterly misleading during an incident.

However if there was no queueing at all (the "No Data" case), we'd get a key with the label "Value", which is a little confusing still.  This change uses the worker label (via Grafana) to add a label to the zero metric vector.  It's still weird when viewing more than one worker at a time on the dashboard, but in the simple case is better.