Increase Postgres int4 alerting threshold
One of the known at-risk tables has crossed the 75% threshold.
We do not want to silence this alert for all tables, and there
is already planned work to address this and other tables
by converting to 8-byte integers.

Here we increase the alerting threshold to 80%, which should
give us a few more weeks at the current growth rate before
we re-alert.  But if the growth rate changes, we will still
retain visibility into that via this alert.

Incident issue: https://gitlab.com/gitlab-com/gl-infra/production/-/issues/5335