Or instead of + to combine apdex and error SLIs
Using or and a label_replace means well make sure that the error SLI
is included when the error-counter is missing (for example when no
errors occurred).

Adding the `sli_kind` label is needed for making sure we keep both
metrics if the first one is present.

The net effect of this that for all groups the budget will look
better. Because in the general case, the error-rate would be much
lower than the apdex-violation rate. This means we're adding a whole
bunch of successful operations.

This does not change the recommendations made based on the error
budgets. The attribution panels correctly take error ratio into
account. So for most groups, apdex is still the biggest contributor to
budget spend.