Add a per-worker queue panel to sidekiq overview
With catchall now mostly using one queue (default), and more to come, while per-queue lengths are useful, it'd also be nice to be able to see the per-worker (i.e. inspect jobs within the default queue).  The data isn't quite as nice because if there are no jobs of a given class the metric goes away entirely, but in incident or debugging scenarios it will be useful