Watch oversized sidekiq jobs being rejected
This adds an ES watch to notify us when a Sidekiq job exceeds 5MB
after compression. This would cause it to be rejected with an
ExceedLimitError.

We don't expect any of these messages to occur. But we add this watch
so we would know about them even if the error rate is too low to
trigger any actual alerts.