Record a 0 error rate for missing failures
This records a 0 for the error rate metrics based on
`sidekiq_jobs_failed_total` when the operation rate exists.

This is similar to how we record the key metrics for puma.

This takes static labels of the recording into account. Avoiding the
inclusion of other rules with the same name but a different component.

The `component` label has been removed from the source aggregation
set for feature category metrics, since that's the place where we add
the label, so we shouldn't aggregate on it. It is always empty in the
source metrics.