Record rates when upscaling, and upscale for Sidekiq
The Sidekiq 6h recording rules are timing out for operation rates. We
can upscale these by taking sum_over_time of the next rule down (1h),
multiplying by the frequency of the recording rule (1m in this case),
then dividing by the interval in question (6h).

This is not exactly equivalent - it applies some smoothing - but it's
better than no recording at all. This also allows us to extend this to
longer burn rates in future.