Generate metric names using format in aggregations
This allows specifying a list of burnRates and a metric format and
generates the metric names for recording rules from that instead of
repeating them for every burn rate.

A specific burnrate can be overriden from the format by explicitly
defining it. For example to exclude a recording, or the have a
diverging recording name.