Record application SLI aggregations
An application SLI consists of a total-counter and a success-counter
coming from the monolith.

This adds an abstraction on top of that, so we only need to add a
definition of the SLI with its labels in a single place to be able to
aggregate this into Thanos

For now, the monolith is only emitting `rails_request_apdex` in this
way. This adds the recordings for that so we can start validating
those in global metrics.