Silence unneeded output from dashboards/test-dashboard.sh
With https://gitlab.com/gitlab-com/runbooks/-/merge_requests/3745
we've made setting the `inherit_errexit` shell option optional.
It's however done in a way, that outputs the option value to STDOUT,
which affects the script output.

As this information is not needed by the user and is fully handled
by the script, we can silence it by redirecting grep's output
to devnull.