Remove unneeded runner alerts
These three alerts were added a long time ago as a way to track
if everything is doing well with the general runners service
and to alert on runner capacity limits.

We have now these covered way better with the Apdex and Saturation
metrics, which trigger alerts by default. There is no need for these
old ones.