Add an SLO recording per component
Adding these in addition to the service ones.

The SLO per service, without a component label is used by
some (deprecated) general service alerts in
`thanos-rules-jsonnet/service-alerts.jsonnet` and for the Aggregated
SLI panels in the service overview dashboard.

The SLO per component is used by the SLI specific panels in the
service overviews.