Generate a service-indicator from application SLI
This allows generating a service-indicator to be used in the service
catalog from an application SLI. The following was necessary:

**Add a description to application SLIs**

This description is used for alerts and dashboards.

**Rate Apdex**

This added a `rateApdex` definition. This allows specifying an
apdex using a success-rate and request-rate rather than a histogram
with tolerated and satisfied values.

The `rateApdex` quacks like a `histogramApdex`:

- `apdexSuccessRateQuery`: The success-rate used in
  `gitlab_component_apdex` recordings
- `apdexWeightQuery`: The weight used in `gitlab_component_apdex`
  recordings
- `apdexQuery`: The ratio used for recording
  `gitlab_component_apdex:ratio`.
- `apdexNumerator`: Used for combining the apdex
- `apdexDenominator`: Used for combining the apdex
- `apdexAttribution`: Used for displaying attribution panels on the
  service overview.
- `supportsReflection`: Used when trying to find sli aggregations for
  panels on the service dashboards.

But since this consists of to rates, it does not support
`histogram_quantile` queries. These would otherwise be used to display
a graph in the alert generated by this SLI, and would be shown in the
latency panels for the apdex detail.