Avoid duplicate SLI recordings
When an Application SLI is added to the service catalog in the
`recordingRuleMetrics` it will generate recordings based on the
defined significant labels. This would be effectively the same
recording as this one, so this skips adding that rule.

Performing this filter allows us to define SLIs and seeing how they
perform in aggregations without adding the SLI to a service.