Add the rails_requests SLI to appropriate services
This adds the `rails_requests` SLI to `web`, `api` and `git`. This SLI
currently only includes an apdex, we could move the `error` component
from `puma` inside it at some point.

A custom `apdexScore` threshold is set for each service, this is based
on the last 7d of the metrics, excluding a single
incident (https://gitlab.com/gitlab-com/gl-infra/production/-/issues/5669)
that caused apdex drops because of a database slowdown.

The `rails_requests` apdex is how many request meet the request
duration target defined in the application. The default target
duration is 1s. Because of that, we need a lower `apdexScore` for each
of these, since requests taking between 1s and 10s are considered
unsatisfactory, instead of tolerated.

This also adds the source metrics to the recording rule registry. So
these can be reused for the gitlab component recordings. This replaces
the old `gitlab_sli_aggregation` recordings.