Use component='' selector for service overview
This will use the `component=''` for service overview panels showing
the default service SLO in those panels. Only the SLIs with a
customized SLO will fall back the correct threshold.

The disadvantage is that when one of the SLIs with an adjusted SLO
dips, but not far enough to alert, the overview row might look like it
should have alerted.

I think this is better, because an `avg(slo)` does not seem right
either.