Add component label to feature_category set
The aggregation set definition should reflect the labels that are
available on the metrics. The source set has the `component` label,
just like the global set does. This corrects that for the
`featureCategorySourceSLIs`.

Filtering out the static labels is an approach copied over from the regular
`serviceLevelIndicator` recording into an aggregation set:
https://gitlab.com/gitlab-com/runbooks/blob/4f94f6a12326243b3f116e0f86812c6b8dc31504/libsonnet/servicemetrics/service_level_indicator_definition.libsonnet#L50

This will be required when we want to use this aggregation set to
record form service level indicators.