Separate teams into it's own yaml
But still combining them into the service catalog.

This will make this file easier to manage once we add all of the stage
groups.

This also means that the generated JSON file is now required where we
used to read the yaml file. This happened in validations for alerts
and services.