Map a team to a stageGroup
This uses the `product_stage_group` value of a team to map it to the
correct stage group from the stage group feature category mapping

The information is then also available on the group object.

This also adds some basic assertions on the validity of the
`product_stage_group` value. These assertions should make the jsonnet
tests fail if we forgot to update the teams.yml when the feature
category mapping changed.

We don't default the name of `product_stage_group` in a team to the
name, because a team could be independent of stage groups. For
example: rapid-action groups.