Record component:feature_category metrics for SLIs
This adds the possiblity to set `fromSourceMetrics` as a
`featureCategory` on the SLI.

Doing so will cause `gitlab:component:feature_category:` type
recording rules to be generated with the `feature_category` as an
aggregation label.

This also adds this for the `rails_requests` component, so we have a
way to feed these metrics into the error budget.