feat(glsh): new `kube` subcommand
Background
---
Currently each engineer that needs to interect with the kubernetes
clusters need to set up `sshuttle` or socks proxy with specific ips, and
remember to change the kubectl context. It seems like everyone has their
own setup and always struggle to keep track of ips and which context
matches with cluster.

Solution
---
Create a new command `glsh kube` which is used to set up the engineers
local machine to interect with our kubernetes clusters.

The commands the engineer has to run are:
```
glsh kube setup
glsh kube use-cluster gstg
```